backend/pyproject.toml
# NOTE: This file was partially generated using AI assistance.
[project]
name = "face-detection-cropping-backend"
version = "0.1.0"
description = "Face detection and cropping API service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "opencv-python>=4.11.0.86",
    "pillow",
    "aiofiles",
    "python-jose[cryptography]",
    "pydantic",
    "pydantic-settings"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.pytest.ini_options]
addopts = "-v -s"
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
